backend/requirements.txt
fastapi
uvicorn
pydantic
pyyaml
netmiko

